refactor: controller 패키지 분리, view 메서드 이름 변경
